Split SkinnyEngineBase into feature traits
