+++ engine/src/main/scala/skinny/engine/async/FutureSupport.scala
@@ -1,5 +1,7 @@
 package skinny.engine.async
 
+import scala.language.postfixOps
+
 import java.util.concurrent.atomic.AtomicBoolean
 import javax.servlet.{ AsyncEvent, AsyncListener }
 
@@ -9,7 +11,6 @@ import skinny.engine.routing.AsyncRoutingDsl
 
 import scala.concurrent.duration._
 import scala.concurrent.{ ExecutionContext, Future }
-import scala.language.postfixOps
 import scala.util.{ Failure, Success }
 
 trait FutureSupport extends SkinnyEngineBase with AsyncRoutingDsl {
@@ -61,7 +62,7 @@ trait FutureSupport extends SkinnyEngineBase with AsyncRoutingDsl {
                     renderHaltException(e)
                   case e =>
                     try {
-                      renderResponse(errorHandler(e))
+                      renderResponse(currentErrorHandler.apply(e))
                     } catch {
                       case e: Throwable =>
                         SkinnyEngineBase.runCallbacks(Failure(e))
@@ -98,7 +99,7 @@ trait FutureSupport extends SkinnyEngineBase with AsyncRoutingDsl {
               case e: HaltException => renderHaltException(e)
               case e =>
                 try {
-                  renderResponse(errorHandler(e))
+                  renderResponse(currentErrorHandler.apply(e))
                 } catch {
                   case e: Throwable =>
                     SkinnyEngineBase.runCallbacks(Failure(e))
