+++ engine/src/main/scala/skinny/engine/base/BeforeAfterDsl.scala
@@ -1,0 +1,26 @@
+package skinny.engine.base
+
+import skinny.engine.RouteTransformer
+import skinny.engine.routing.Route
+
+trait BeforeAfterDsl { self: RouteRegistryAccessor =>
+
+  /**
+   * Adds a filter to run before the route.  The filter only runs if each
+   * routeMatcher returns Some.  If the routeMatchers list is empty, the
+   * filter runs for all routes.
+   */
+  def before(transformers: RouteTransformer*)(fun: => Any): Unit = {
+    routes.appendBeforeFilter(Route(transformers, () => fun))
+  }
+
+  /**
+   * Adds a filter to run after the route.  The filter only runs if each
+   * routeMatcher returns Some.  If the routeMatchers list is empty, the
+   * filter runs for all routes.
+   */
+  def after(transformers: RouteTransformer*)(fun: => Any): Unit = {
+    routes.appendAfterFilter(Route(transformers, () => fun))
+  }
+
+}
