+++ engine/src/main/scala/skinny/engine/base/RouteRegistryAccessor.scala
@@ -1,0 +1,12 @@
+package skinny.engine.base
+
+import skinny.engine.routing.RouteRegistry
+
+trait RouteRegistryAccessor {
+
+  /**
+   * The routes registered in this kernel.
+   */
+  lazy val routes: RouteRegistry = new RouteRegistry
+
+}
