--- engine/src/main/scala/skinny/engine/cookie/CookieSupport.scala
@@ -1,9 +1,0 @@
-package skinny.engine.cookie
-
-object CookieSupport {
-
-  val SweetCookiesKey = "skinny.engine.SweetCookies"
-
-  val CookieOptionsKey = "skinny.engine.CookieOptions"
-
-}
